+++ src/Brmble.Client/Services/Voice/AudioManager.cs
@@ -115,97 +115,6 @@ internal static class AudioLog
     }
 }
 
-internal static class Win32RawInput
-{
-    public const uint WM_INPUT = 0x00FF;
-    public const uint RID_INPUT = 0x10000003;
-    public const uint RIDEV_REMOVE = 0x00000001;    // Remove device registration
-    public const uint RIDEV_INPUTSINK = 0x00000100; // Receive input even when window not focused
-    public const uint RIDEV_NOLEGACY = 0x00000030;  // Don't receive legacy messages (blocks key)
-    public const uint RIM_TYPEKEYBOARD = 1;         // Matches RAWINPUTHEADER.dwType for keyboards
-    public const ushort HID_USAGE_PAGE_GENERIC = 0x01;
-    public const ushort HID_USAGE_GENERIC_KEYBOARD = 0x06;
-    public const ushort HID_USAGE_GENERIC_MOUSE = 0x02;
-
-    public const int WH_MOUSE_LL = 14;
-    public const int WH_KEYBOARD_LL = 13;
-    public const int WM_KEYDOWN = 0x0100;
-    public const int WM_KEYUP = 0x0101;
-    public const int WM_SYSKEYDOWN = 0x0104;
-    public const int WM_SYSKEYUP = 0x0105;
-    public const int WM_LBUTTONDOWN = 0x0201;
-    public const int WM_RBUTTONDOWN = 0x0204;
-    public const int WM_MBUTTONDOWN = 0x0207;
-    public const int WM_XBUTTONDOWN = 0x020B;
-    public const int WM_LBUTTONUP = 0x0202;
-    public const int WM_RBUTTONUP = 0x0205;
-    public const int WM_MBUTTONUP = 0x0208;
-    public const int WM_XBUTTONUP = 0x020C;
-
-    public const int XBUTTON1 = 1;
-    public const int XBUTTON2 = 2;
-
-    [StructLayout(LayoutKind.Sequential)]
-    public struct MSLLHOOKSTRUCT
-    {
-        public int ptX;
-        public int ptY;
-        public int mouseData;
-        public int flags;
-        public int time;
-        public IntPtr dwExtraInfo;
-    }
-
-    [StructLayout(LayoutKind.Sequential)]
-    public struct RAWINPUTHEADER
-    {
-        public uint dwType;
-        public uint dwSize;
-        public IntPtr hDevice;
-        public IntPtr wParam;
-    }
-
-    [StructLayout(LayoutKind.Sequential)]
-    public struct RAWKEYBOARD
-    {
-        public ushort MakeCode;
-        public ushort Flags;
-        public ushort Reserved;
-        public ushort VKey;
-        public uint Message;
-        public uint ExtraInformation;
-    }
-
-    [DllImport("user32.dll", CharSet = CharSet.Auto, SetLastError = true)]
-    public static extern int GetRawInputData(IntPtr hRawInput, uint uiCommand, IntPtr pData, ref uint pcbSize, uint cbSizeHeader);
-
-    [DllImport("user32.dll", CharSet = CharSet.Auto, SetLastError = true)]
-    public static extern bool RegisterRawInputDevices(RAWINPUTDEVICE[] pRawInputDevices, uint uiNumDevices, uint cbSize);
-
-    public delegate IntPtr LowLevelMouseProc(int nCode, IntPtr wParam, IntPtr lParam);
-
-    [DllImport("user32.dll", CharSet = CharSet.Auto, SetLastError = true)]
-    public static extern IntPtr SetWindowsHookEx(int idHook, LowLevelMouseProc lpfn, IntPtr hMod, uint dwThreadId);
-
-    [DllImport("user32.dll", CharSet = CharSet.Auto, SetLastError = true)]
-    public static extern bool UnhookWindowsHookEx(IntPtr hhk);
-
-    [DllImport("user32.dll", CharSet = CharSet.Auto, SetLastError = true)]
-    public static extern IntPtr CallNextHookEx(IntPtr hhk, int nCode, IntPtr wParam, IntPtr lParam);
-
-    [DllImport("kernel32.dll", CharSet = CharSet.Auto, SetLastError = true)]
-    public static extern IntPtr GetModuleHandle(string? lpModuleName);
-
-    [StructLayout(LayoutKind.Sequential)]
-    public struct RAWINPUTDEVICE
-    {
-        public ushort usUsagePage;
-        public ushort usUsage;
-        public uint dwFlags;
-        public IntPtr hwndTarget;
-    }
-}
-
 /// <summary>
 /// Manages audio I/O: mic capture via EncodePipeline, per-user speaker
 /// playback via UserAudioPipeline, mute/deafen state, and speaking detection.
